requirements.txt
dajngo
django-celery-beat
reportlab